% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_support_rule}
\alias{get_support_rule}
\title{Replicate-support decision rule}
\usage{
get_support_rule(
  n_valid_target,
  supporting_reps,
  min_support_when_three_valid = 2,
  min_support_when_two_valid = 2
)
}
\arguments{
\item{n_valid_target}{number of valid target replicates (2 or 3)}

\item{supporting_reps}{number of supporting replicates}

\item{min_support_when_three_valid}{required supports for 3 valid reps}

\item{min_support_when_two_valid}{required supports for 2 valid reps}
}
\value{
logical(1)
}
\description{
Requires at least `min_support_when_three_valid` supporting
replicates when 3 valid replicates exist, and `min_support_when_two_valid`
when 2 exist.
}
\keyword{internal}
