% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{read_input_data}
\alias{read_input_data}
\title{Read the annotated MaxQuant Excel workbook}
\usage{
read_input_data(input_file, sheet_name = 1)
}
\arguments{
\item{input_file}{path to the .xlsx workbook}

\item{sheet_name}{sheet index or name (default 1)}
}
\value{
tibble with raw data
}
\description{
Read the annotated MaxQuant Excel workbook
}
