% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{save_analysis_plots}
\alias{save_analysis_plots}
\title{Save analysis plots as PNG files}
\usage{
save_analysis_plots(plots, plots_dir, width = 9, height = 6, dpi = 200)
}
\arguments{
\item{plots}{named list of ggplot objects}

\item{plots_dir}{directory to write PNGs into (created)}

\item{width, height, dpi}{passed to ggplot2::ggsave}
}
\value{
invisibly, the plots directory
}
\description{
Save analysis plots as PNG files
}
